% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dropdown.R
\name{dropdown_UI}
\alias{dropdown_UI}
\alias{dropdown}
\title{Semantic dropdown menu module}
\usage{
dropdown_UI(id, choices, def_text)

dropdown(id)
}
\arguments{
\item{id}{namespace id}

\item{choices}{character, vector of possible selections}

\item{def_text}{character, a descriptor of the dropdown field}
}
\description{
Generates the UI and server logic for a Semantic/Fomantic 
dropdown menu. The server is quite simple, merely returning the 
value selected in the UI to be used downstream
}
