% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vessel_card.R
\name{vessel_card}
\alias{vessel_card}
\title{vessel_card}
\usage{
vessel_card(vdata)
}
\arguments{
\item{vdata}{a data frame with vessel data}
}
\value{
the HTML to make semantic card
}
\description{
A function to create the card with the vessel information
}
\examples{
if(interactive()) {vessel_card(distances[1,])}
}
