% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vessel_map.R
\name{vessel_map}
\alias{vessel_map}
\title{vessel_map}
\usage{
vessel_map(vdata)
}
\arguments{
\item{vdata}{a data frame with vessel data}
}
\value{
a leaflet map with two points for the start and end of the trip
}
\description{
A function to create a leaflet map
}
\examples{
if(interactive()) {vessel_map(distances[1,])}
}
